requirements.txt
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.1.0
beartype==0.16.4
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
denoising-diffusion-pytorch==1.9.4
diffusers==0.23.1
docopt==0.6.2
einops==0.7.0
ema-pytorch==0.3.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1700579780973/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss==1.7.4
faiss-cpu==1.7.4
filelock==3.13.1
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
geographiclib==2.0
geopy==2.4.1
huggingface-hub==0.19.4
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.0
importlib-metadata==6.8.0
invisible-watermark==0.2.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1698244021190/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1700832787564/work
ipywidgets==8.1.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.3.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1698673647019/work
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
matplotlib==3.8.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mlxtend==0.23.1
multidict==6.0.4
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.2.1
nibabel==5.1.0
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=fc2732718bc9e06a7b702492cb4f5afffe9671083930452d894377bf563464a3
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///croot/pillow_1696580024257/work
pipreqs==0.4.13
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1699715570510/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1699963054032/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.4.3
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-fid==0.3.0
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
SimpleITK==2.3.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
# Editable install with no version control (stable-diffusion==0.0.1)
-e /home/zhangsenpeng/stable_diffusion/stablediffusion
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
staticmap==0.5.7
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.15.0
torch==1.12.1
torch-summary==1.4.5
torchaudio==0.12.1
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.13.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698671135544/work
transformers==4.35.2
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1698257533958/work
utm==0.7.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1700607916581/work
widgetsnbextension==4.0.9
xformers==0.0.22
yarg==0.1.9
yarl==1.9.3
zipp==3.17.0
